Move some of the definitions in src/data
## Description

The placement of some of these definitions was pretty haphazard as I
was initially doing the TS conversion. This change cleans it up a bit
with the benefit of hindsight.

The new rule is that every .ts file directly in `src/data` should
correspond directly to one or more static data files, and the .ts file
should load its corresponding static data file(s). This meant that I
moved `items.ts` and `states.ts`, which just define enums for use by
other files, into the new subdir `types`.

Additionally, it bothered me that `state_incentives.ts` was importing
things from `ira_incentives.ts`. I moved the shared definitions into
their own files in `types`.

## Test Plan

Unit tests pass. This is all just code motion, no functionality changes.
