+++ src/data/types/owner-status.ts
@@ -1,0 +1,4 @@
+export enum OwnerStatus {
+  Homeowner = 'homeowner',
+  Renter = 'renter',
+}
